requirements.txt
requests
kafka-python
apache-airflow
pyspark
psycopg2-binary
prometheus-client